% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{ggvisForSymbol}
\alias{ggvisForSymbol}
\title{demonstrate ggplot-based gene model concept}
\usage{
ggvisForSymbol(sym, resource = EnsDb.Hsapiens.v79::EnsDb.Hsapiens.v79,
  columnsKept = c("gene_id", "tx_id"), yval = 1, arrmm = 1.5,
  viewtype = "transcripts", ...)
}
\arguments{
\item{sym}{character(1) gene symbol}

\item{resource}{an instance of \code{\link{EnsDb-class}}}

\item{yval}{numeric(1) vertical position of segments}

\item{arrmm}{numeric(1) number of millimeters for arrowhead segments}

\item{viewtype}{character(1) either "exons" or "transcripts"}

\item{\dots}{passed to GenomicFeatures::exons}
}
\value{
ggplot instance
}
\description{
demonstrate ggplot-based gene model concept
}
\note{
It is known that arrow heads do not propagate to ggplotly
displays as of Aug 2018.
}
\examples{
requireNamespace("EnsDb.Hsapiens.v75")
edb = EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75
ggvisForSymbol("ORMDL3", arrmm=1.3, resource=edb) + ylab(" ") + 
      theme(axis.text.y = element_blank()) + ylim(-1,3)
}
